Generates noises based off array of soundBuffers. Added button to generate and play sounds each time.

+++ AudioSettings.h
@@ -4,6 +4,6 @@
 
 namespace AudioSettings
 {
-  static const int64_t AUDIO_SAMPLE_RATE = 48000;
+  static const uint32_t AUDIO_SAMPLE_RATE = 48000;
   //static const int64_t AUDIO_SAMPLE_RATE = 96000;
 }
